IGNITE-23436 Do not use ByteUtils#toBytes to persist objects related to CMG (#4562)

+++ modules/partition-replicator/src/main/java/org/apache/ignite/internal/partition/replicator/marshaller/PartitionCommandsMarshallerImpl.java
@@ -64,6 +64,6 @@ public class PartitionCommandsMarshallerImpl extends OptimizedMarshaller impleme
      */
     @Override
     public int readRequiredCatalogVersion(ByteBuffer raw) {
-        return VarIntUtils.readVarInt(raw);
+        return (int) VarIntUtils.readVarInt(raw);
     }
 }
